Adjust to bindable QQuickItem properties
This prepares for QQuickItem's properties becoming bindable, changing
the type of QQuickItemPrivate::{x, y, width, height}. It adds explicit
casts/types where implicit conversion or type deduction would fail after
the conversion.

+++ src/quicktemplates2/qquicksplitview.cpp
@@ -338,8 +338,8 @@ void QQuickSplitViewPrivate::layoutResizeSplitItems(qreal &usedWidth, qreal &use
         if (isBeingResized) {
             // Don't let the mouse go past either edge of the SplitView.
             const qreal clampedMousePos = horizontal
-                ? qBound(qreal(0.0), m_mousePos.x(), width)
-                : qBound(qreal(0.0), m_mousePos.y(), height);
+                ? qBound(qreal(0.0), m_mousePos.x(), qreal(width))
+                : qBound(qreal(0.0), m_mousePos.y(), qreal(height));
 
             // We also need to ensure that the item's edge doesn't go too far
             // out and hence give the item more space than is available.
